requirements.txt
streamlit
Pillow
tensorflow
numpy

